[feat] firebase Card create, read 구현
