+++ package.json
@@ -29,6 +29,7 @@
     "react-kakao-maps-sdk": "^1.1.6",
     "react-responsive": "^9.0.2",
     "react-router-dom": "6.8.2",
+    "react-select": "^5.7.0",
     "recoil": "0.7.7",
     "swiper": "9.1.0"
   },
