+++ src/components/Modal/Modal.module.scss
@@ -103,3 +103,4 @@
 .signupButton {
   margin-top: 35px;
 }
+
